requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
gradio>=4.0.0
pydantic>=1.8.2
python-multipart>=0.0.5
requests>=2.31.0
schedule==1.2.0
cryptography==41.0.0